get rid of dependency cycle error
