+++ gameEngine/specials.js
@@ -5,26 +5,9 @@
 const {
     card: cardConst,
     target: targetConst,
-    action,
 } = require('../src/constants');
 
 const { getRandomUpTo } = require('../gameTerminal/randomFunc');
-
-const { giveCardsTo } = require('./game');
-
-const {
-    getActivePlayer,
-    getInActivePlayer,
-    lastActionChange,
-    moveCardGraveyard,
-    attackShield,
-    changeTurn,
-} = require('./actions');
-// const { getInActivePlayer } = require('./actions');
-// const { lastActionChange } = require('./actions');
-// const { moveCardGraveyard } = require('./actions');
-// const { attackShield } = require('./actions');
-// const { changeTurn } = require('./actions');
 
 // helper function to get random index for player's cards in hand
 function getRandomIndexes(cardsLength) {
@@ -67,33 +50,6 @@ function bowArrow(player, opponent) {
     }
 }
 
-// function forestMushroom accepts opponent
-// and if opponent get 60% chance, then his/ her cards in hand
-// get panic: true property except one random card
-// so at the begginig of opponent action he can play only this card
-function forestMushroom(game, opponent) {
-    // console.log('We ate forest mushrooms!!', opponent.hero);
-    const chance = getRandomUpTo(10, 'chanceforestMushroom');
-    const opponentCards = Object.values(opponent.hand);
-    if (chance <= 6) {
-        for (let i = 0; i < Object.keys(opponent.hand).length; i++) {
-            opponentCards[i].panic = true;
-        }
-        Object.keys(opponent.item).length !== 0
-            ? (Object.values(opponent.item)[0].panic = true)
-            : null;
-        const index = getRandomUpTo(opponentCards.length, 'indexMushroom');
-        opponentCards[index].panic = false;
-        if (opponentCards[index].disabled === true) {
-            opponentCards[index].panic = false;
-            opponentCards[index].disabled = false;
-        }
-        // after player's act we change lastAction property of the game
-        lastActionChange(game, action.CHAINS);
-    }
-    return opponentCards;
-}
-
 // function to remove panic true (for forestMushroom card)
 // from player's hand cards
 function removePanic(player) {
@@ -103,60 +59,6 @@ function removePanic(player) {
         delete Object.values(player.item)[0].panic;
     }
     return playerCards;
-}
-
-// function checks whether opponent item is not empty
-// and whether opponent has magicTree card in item
-// if so - function change turn runs after moveCounter === 1, active layer becomes inactive etc.
-function magicTree(game) {
-    let itemId;
-    const pActive = getActivePlayer(game);
-    const pInactive = getInActivePlayer(game);
-    const itemKey = Object.keys(pInactive.item)[0];
-    itemKey ? (itemId = pInactive.item[itemKey].id) : null;
-    if (pActive.moveCounter === 1 && itemId === cardConst.MAGICTREECARD) {
-        giveCardsTo(pActive);
-        changeTurn(game);
-    }
-}
-
-// TODO: rename to magicMirror/plateMail or not
-// function to reflect half of the damage (or round down to integer) for magicMirror card
-// if active card has 1 attack points then each player gets 1 pnt damage
-// or in case of plateMail func reflect 1 point back to player
-// if active card has 1 attack points then no damage to both players
-function reflect(opponent, player, points) {
-    // console.log('We are in reflect function!', player);
-    const itemCard = Object.values(opponent.item)[0].id;
-    let damage;
-    if (itemCard === cardConst.MAGICMIRRORCARD) {
-        points === 1 ? (damage = 1) : (damage = Math.floor(points / 2));
-        opponent.health.current -= damage;
-    }
-    if (itemCard === cardConst.PLATEMAILCARD) {
-        points === 1 ? (damage = 0) : (damage = 1);
-        damage === 0
-            ? opponent.health.current
-            : (opponent.health.current -= points - damage);
-    }
-    if (
-        Object.keys(player.item).length === 0
-    || Object.values(player.item)[0].category !== cardConst.SHIELDCARD
-    ) {
-        player.health.current -= damage;
-    }
-    if (
-        Object.keys(player.item).length === 1
-    && Object.values(player.item)[0].category === cardConst.SHIELDCARD
-    ) {
-        attackShield(player, Object.keys(player.item)[0], damage);
-    }
-    if (player.health.current < 0) {
-        player.health.current = 0;
-    }
-    if (opponent.health.current < 0) {
-        opponent.health.current = 0;
-    }
 }
 
 function cardIncreaseHealth(players) {
@@ -278,41 +180,9 @@ function removeDisable(player) {
     return playerCards;
 }
 
-// TODO: rename to Skull Lantern
-// this function runs when player attacks with the card Skull Lantern which category is attackItems
-// it accept players and checks all item category cards from both players item holders and hands
-// and move such cards to players grave yards
-function attackItems(players) {
-    players.forEach((p) => {
-        if (Object.keys(p.item).length !== 0) {
-            const itemCard = Object.values(p.item)[0];
-            // we reset item card's health points to initial points
-            itemCard.healthCurrent = itemCard.health;
-            // we move any item card to graveyard
-            moveCardGraveyard(p, Object.keys(p.item)[0], cardConst.ITEMCARD);
-        }
-        // we check whether each player hand is not empty
-        if (Object.keys(p.hand).length !== 0) {
-            // and for each card in hand with type item
-            for (const cardIndex in p.hand) {
-                if (p.hand[cardIndex].type === cardConst.ITEMCARD) {
-                    // we reset item card's points to initial points
-                    p.hand[cardIndex].healthCurrent = p.hand[cardIndex].health;
-                    // we move any item card to graveyard
-                    moveCardGraveyard(p, cardIndex);
-                }
-            }
-        }
-    });
-    return { players };
-}
-
 module.exports = {
     bowArrow,
-    forestMushroom,
     removePanic,
-    magicTree,
-    reflect,
     waterCard,
     malachiteBox,
     turningHand,
@@ -320,5 +190,4 @@ module.exports = {
     deleteCardsShown,
     disableCards,
     removeDisable,
-    attackItems,
 };
